Smart_Product_Pricing/requirements.txt
pandas==2.2.2
numpy==2.1.2
scikit-learn==1.5.2
torch==2.4.1
torchvision==0.19.1
matplotlib==3.9.2
seaborn==0.13.2
tqdm==4.66.5
jupyterlab==4.2.5
jinja2==3.1.4
sentence-transformers==3.2.1
transformers==4.45.2
aiohttp==3.10.10
aiofiles==24.1.0
Pillow==10.4.0

